move diff logic to wrangler src utils
